Internal: move Blockifier specific code to the correct module
