+++ tests/integration/common/state.rs
@@ -5,7 +5,7 @@ use blockifier::context::BlockContext;
 use blockifier::state::cached_state::CachedState;
 use blockifier::test_utils::dict_state_reader::DictStateReader;
 use blockifier::test_utils::BALANCE;
-use cairo_lang_starknet_classes::casm_contract_class::{CasmContractClass, StarknetSierraCompilationError};
+use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
 use cairo_lang_starknet_classes::contract_class::ContractClass;
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
@@ -22,7 +22,7 @@ use starknet_os::storage::dict_storage::DictStorage;
 use starknet_os::storage::storage::{FactFetchingContext, HashFunctionType, StorageError};
 use starknet_os::storage::storage_utils::{compiled_contract_class_cl2vm, deprecated_contract_class_api2vm};
 
-use super::blockifier_contracts::{get_feature_sierra_contract_class, load_cairo0_feature_contract};
+use super::blockifier_contracts::{load_cairo0_feature_contract, load_cairo1_contract};
 use crate::common::block_context;
 use crate::common::blockifier_contracts::get_deprecated_erc20_contract_class;
 
@@ -100,25 +100,6 @@ pub struct FeeContracts {
     pub erc20_contract: DeprecatedCompiledClass,
     pub eth_fee_token_address: ContractAddress,
     pub strk_fee_token_address: ContractAddress,
-}
-
-/// Compiles a Sierra class to CASM.
-fn compile_sierra_contract_class(
-    sierra_contract_class: ContractClass,
-) -> Result<CasmContractClass, StarknetSierraCompilationError> {
-    // Values taken from the defaults of `starknet-sierra-compile`, see here:
-    // https://github.com/starkware-libs/cairo/blob/main/crates/bin/starknet-sierra-compile/src/main.rs
-    let add_pythonic_hints = false;
-    let max_bytecode_size = 180000;
-    CasmContractClass::from_contract_class(sierra_contract_class, add_pythonic_hints, max_bytecode_size)
-}
-
-/// Helper to load a Cairo1 contract class.
-pub fn load_cairo1_contract(name: &str) -> (String, ContractClass, CasmContractClass) {
-    let sierra_contract_class = get_feature_sierra_contract_class(name);
-    let casm_contract_class = compile_sierra_contract_class(sierra_contract_class.clone())
-        .unwrap_or_else(|e| panic!("Failed to compile Sierra contract {}: {}", name, e));
-    (name.to_string(), sierra_contract_class, casm_contract_class)
 }
 
 /// Configures the logging for integration tests.
